Improve mockr project search speed
As a hackathon project, we use some smarter search techniques to hide
and show fewer list elements when each search is performed. For each
keystroke, this narrows down or broadens the existing list instead of
hiding and showing everything.

This is motivated by the slowdown in speed with large quantities of
projects.

Change-Id: I6f5e916c2f2ac1a8a4e339e69116a513a23ac0da
Reviewed-on: https://gerrit.causes.com/18733
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>
Tested-by: Tom Dooner <tom.dooner@causes.com>