Allow sorting mocks by designer
Adds a field to users to specify the designers and adds a field to
change this on the user edit page.
Stops the recent mocks from being split into groups of 3, so that they
can act more responsive, and puts them all inside of one list.
Puts a list of designers on the sidebar, that when clicked, filters
the mocks to only show mocks by that designer.

Conflicts:

    public/javascripts/home/index.js

Change-Id: I9c4a51237e70ad3d8960c3978d160cd64a8d71c0
Reviewed-on: https://gerrit.causes.com/18763
Tested-by: Aiden Scandella <aiden@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>