Allows sorting mocks be designer.
Adds a field to users to specify the designers and adds a field to
change this on the user edit page.
Stops the recent mocks from being split into groups of 3, so that they
can act more responsive, and puts them all inside of one list.
Puts a list of designers on the sidebar, that when clicked, filters
the mocks to only show mocks by that designer.