List three most recently viewed mocks on homepage
Noticed that mock views were being tracked, but not used for anything.
Because most users will likely be going to Mockr to look at the same
mocks over and over, a list of recently viewed mocks makes it easier
to find what the user is looking for.

To avoid confusion between "Recently Viewed" and "Recent Mocks" (the list
of latest mocks created), "Recent Mocks" was renamed to "New Mocks".

Conflicts:
	app/views/home/index.html.erb

Change-Id: Id57e5746b3b7cc7ac43baf581d12a86549221525
Reviewed-on: https://gerrit.causes.com/18764
Tested-by: Aiden Scandella <aiden@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>