Add 'Latest' link to navbar
This commit adds a 'Latest' link next to a given mock that can be used
to always link to/reference the latest version of a stack of mocks.
This is useful for, e.g., linking to mocks in documents without ever
having to worry about them going out of date.

Change-Id: I04aa040971f9ee0b0d337d547cec4cdd295ca0d8
Reviewed-on: https://gerrit.causes.com/20183
Tested-by: Tony Wooster <tony.wooster@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>